+++ test/test_ai_toolkit.rb
@@ -75,7 +75,7 @@ class TestAiToolkit < Minitest::Test
     assert_equal "end_turn", resp.stop_reason
     assert_equal "hi", resp.messages.first[:content]
     assert_equal 1, resp.results.length
-    assert_instance_of AiToolkit::MessageResult, resp.results.first
+    assert_instance_of AiToolkit::Results::MessageResult, resp.results.first
   end
   # rubocop:enable Metrics/MethodLength, Metrics/AbcSize
 
@@ -99,9 +99,9 @@ class TestAiToolkit < Minitest::Test
     assert_equal "end_turn", resp.stop_reason
     assert_equal "done", resp.messages.first[:content]
     assert_equal 3, resp.results.length
-    assert_instance_of AiToolkit::ToolRequest, resp.results[0]
-    assert_instance_of AiToolkit::ToolResponse, resp.results[1]
-    assert_instance_of AiToolkit::MessageResult, resp.results[2]
+    assert_instance_of AiToolkit::Results::ToolRequest, resp.results[0]
+    assert_instance_of AiToolkit::Results::ToolResponse, resp.results[1]
+    assert_instance_of AiToolkit::Results::MessageResult, resp.results[2]
   end
   # rubocop:enable Metrics/MethodLength, Metrics/AbcSize
 
@@ -125,9 +125,9 @@ class TestAiToolkit < Minitest::Test
 
     assert_equal "tool_use", resp.stop_reason
     assert_equal 3, resp.results.length
-    assert_instance_of AiToolkit::ToolRequest, resp.results[0]
-    assert_instance_of AiToolkit::ToolResponse, resp.results[1]
-    assert_instance_of AiToolkit::ToolRequest, resp.results[2]
+    assert_instance_of AiToolkit::Results::ToolRequest, resp.results[0]
+    assert_instance_of AiToolkit::Results::ToolResponse, resp.results[1]
+    assert_instance_of AiToolkit::Results::ToolRequest, resp.results[2]
   end
   # rubocop:enable Metrics/MethodLength, Metrics/AbcSize
 
@@ -226,7 +226,7 @@ class TestAiToolkit < Minitest::Test
     assert_equal 1, provider.call_count
     assert_equal "tool_use", resp.stop_reason
     assert_equal 1, resp.results.length
-    assert_instance_of AiToolkit::ToolRequest, resp.results.first
+    assert_instance_of AiToolkit::Results::ToolRequest, resp.results.first
   end
   # rubocop:enable Metrics/MethodLength, Metrics/AbcSize
 
@@ -266,8 +266,8 @@ class TestAiToolkit < Minitest::Test
 
     assert_equal "tool_stop", resp.stop_reason
     assert_equal 2, resp.results.length
-    assert_instance_of AiToolkit::ToolRequest, resp.results[0]
-    assert_instance_of AiToolkit::ToolResponse, resp.results[1]
+    assert_instance_of AiToolkit::Results::ToolRequest, resp.results[0]
+    assert_instance_of AiToolkit::Results::ToolResponse, resp.results[1]
   end
   # rubocop:enable Metrics/MethodLength, Metrics/AbcSize
 
@@ -294,7 +294,7 @@ class TestAiToolkit < Minitest::Test
     assert_equal 3, resp.results.length
     expected = %w[one two done]
     resp.results.each_with_index do |r, i|
-      assert_instance_of AiToolkit::MessageResult, r
+      assert_instance_of AiToolkit::Results::MessageResult, r
       assert_equal expected[i], r.content
     end
   end